Fixed PHP Warning when looking for Resque
If we're using Resque - we are almost guaranteed to load it ourselves in order to configure it. Since `class_exists()` tries to autoload by default, this raises a PHP Warning every time an exception is handled. Setting `$autoload=false` on the parameter list makes sure it only tests for already existing Resque, not attempt to include it as well.